Add top level hash function
